requirements.txt

python-dotenv
openai
python-docx
PyPDF2
temporalio

flask
fastapi
python-multipart
markdown
beautifulsoup4

sqlalchemy>=2.0
